+++ src/Maker/Resources/skeleton/Category.tpl.php
@@ -1,0 +1,11 @@
+<?= "<?php\n" ?>
+
+namespace <?= $namespace ?>;
+
+<?= $use_statements ?>
+
+enum <?= $class_name ?> :int implements CategoryInterface 
+{
+    case FOO = 1;
+    case BAR = 2;
+}