+++ src/PsysOrderInvoiceBundle.php
@@ -2,9 +2,6 @@
 
 namespace Psys\OrderInvoiceBundle;
 
-use Psys\OrderInvoiceBundle\DependencyInjection\Compiler\ResolveTargetEntityPass;
-use Symfony\Component\Config\Definition\Configurator\DefinitionConfigurator;
-use Symfony\Component\DependencyInjection\Compiler\PassConfig;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 use Symfony\Component\HttpKernel\Bundle\AbstractBundle;