+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 All notable changes to this project are documented in this file.
 
 
+## [1.2.1] - 2025-07-18
+### Added
+Commands and makers for easy configuration and usage:
+- `symfony console oib:configure`
+- `symfony console make:oib:category`
+- `symfony console make:oib:cron_controller`
+
+### Changed
+
+### Fixed
+
+
 ## [1.2.0] - 2025-07-07
 ### Added